requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.0.0
python-multipart>=0.0.18
PyMuPDF>=1.25.0
ollama>=0.4.0
redis>=5.0.0
httpx>=0.28.0
numpy>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
